Added working gripper
