+++ brain/robot.hh
@@ -28,6 +28,7 @@ class Robot {
     void done();
 
     void set_vel(double lvel, double rvel);
+    void set_arm_ang(double ang);
 
     void on_scan(ConstSonarStampedPtr &msg);
     void on_frame(ConstImageStampedPtr &msg);
@@ -38,6 +39,7 @@ class Robot {
 
     gazebo::transport::NodePtr node;
     gazebo::transport::PublisherPtr vel_pub;
+    gazebo::transport::PublisherPtr arm_pub;
     gazebo::transport::SubscriberPtr scan_sub;
     gazebo::transport::SubscriberPtr frame_sub;
     gazebo::transport::SubscriberPtr pose_sub;
