+++ models/tankbot/model.sdf
@@ -46,11 +46,11 @@
         <mass>1</mass>
         <inertia><ixx>0.166667</ixx><ixy>0</ixy><ixz>0</ixz><iyy>0.166667</iyy><iyz>0</iyz><izz>0.166667</izz></inertia>
       </inertial>
-      <pose>0.14650000000000002 0.108 0.135 0 0 0</pose>
+      <pose>0.14650000000000002 -0.108 0.135 0 0 0</pose>
       <visual name='visual'>
         <geometry><box><size>0.35 0.05 0.05</size></box></geometry>
         <material>
-          <script><uri>file://media/materials/scripts/gazebo.material</uri><name>Gazebo/Orange</name></script>
+          <script><uri>file://media/materials/scripts/gazebo.material</uri><name>Gazebo/Grey</name></script>
         </material>
       </visual>
       <collision name='collision'>
@@ -63,12 +63,12 @@
         <inertia><ixx>0.166667</ixx><ixy>0</ixy><ixz>0</ixz><iyy>0.166667</iyy><iyz>0</iyz><izz>0.166667</izz></inertia>
       </inertial>
       <pose>0.40811165235168156
-        0.019611652351681552
-        0.135 0 0 -0.7853981633974483</pose>
+        -0.019611652351681552
+        0.135 0 0 0.7853981633974483</pose>
       <visual name='visual'>
         <geometry><box><size>0.25 0.05 0.05</size></box></geometry>
         <material>
-          <script><uri>file://media/materials/scripts/gazebo.material</uri><name>Gazebo/Orange</name></script>
+          <script><uri>file://media/materials/scripts/gazebo.material</uri><name>Gazebo/Black</name></script>
         </material>
       </visual>
       <collision name='collision'>
@@ -82,8 +82,8 @@
         <inertia><ixx>0.166667</ixx><ixy>0</ixy><ixz>0</ixz><iyy>0.166667</iyy><iyz>0</iyz><izz>0.166667</izz></inertia>
       </inertial>
       <pose>-0.02849999999999997
-        0.108
-        0.135 0 0 -0.7853981633974483</pose>
+        -0.108
+        0.135 0 0 0.7853981633974483</pose>
       <visual name='visual'>
         <geometry><cylinder><radius>0.03</radius><length>0.06</length></cylinder></geometry>
         <material>
@@ -116,11 +116,11 @@
         <mass>1</mass>
         <inertia><ixx>0.166667</ixx><ixy>0</ixy><ixz>0</ixz><iyy>0.166667</iyy><iyz>0</iyz><izz>0.166667</izz></inertia>
       </inertial>
-      <pose>0.14650000000000002 -0.108 0.135 0 0 0</pose>
+      <pose>0.14650000000000002 0.108 0.135 0 0 0</pose>
       <visual name='visual'>
         <geometry><box><size>0.35 0.05 0.05</size></box></geometry>
         <material>
-          <script><uri>file://media/materials/scripts/gazebo.material</uri><name>Gazebo/Orange</name></script>
+          <script><uri>file://media/materials/scripts/gazebo.material</uri><name>Gazebo/Grey</name></script>
         </material>
       </visual>
       <collision name='collision'>
@@ -133,12 +133,12 @@
         <inertia><ixx>0.166667</ixx><ixy>0</ixy><ixz>0</ixz><iyy>0.166667</iyy><iyz>0</iyz><izz>0.166667</izz></inertia>
       </inertial>
       <pose>0.40811165235168156
-        -0.019611652351681552
-        0.135 0 0 0.7853981633974483</pose>
+        0.019611652351681552
+        0.135 0 0 -0.7853981633974483</pose>
       <visual name='visual'>
         <geometry><box><size>0.25 0.05 0.05</size></box></geometry>
         <material>
-          <script><uri>file://media/materials/scripts/gazebo.material</uri><name>Gazebo/Orange</name></script>
+          <script><uri>file://media/materials/scripts/gazebo.material</uri><name>Gazebo/Black</name></script>
         </material>
       </visual>
       <collision name='collision'>
@@ -152,8 +152,8 @@
         <inertia><ixx>0.166667</ixx><ixy>0</ixy><ixz>0</ixz><iyy>0.166667</iyy><iyz>0</iyz><izz>0.166667</izz></inertia>
       </inertial>
       <pose>-0.02849999999999997
-        -0.108
-        0.135 0 0 0.7853981633974483</pose>
+        0.108
+        0.135 0 0 -0.7853981633974483</pose>
       <visual name='visual'>
         <geometry><cylinder><radius>0.03</radius><length>0.06</length></cylinder></geometry>
         <material>
